EC2-001: Fixed paramiko bug

+++ ec2-build.py
@@ -1,5 +1,4 @@
 #!/usr/bin/python
-
 ###### Script will be used to spin up EC2 instances. In this example, I will be 
 ###### using AWS Free Tier account for these examples.
 
@@ -59,13 +58,14 @@ def build(ctx,ami,instancetype,vpc,isweb,subnet,key):
 
     if isweb:
         log.info("Spinning up Web Server")
+        log.info("Passing key: " + key)
         try:
             ec2 = connection() 
             if ec2:
                 log.info("AWS Session Created") 
                 log.info("Gathering VPC ID...")
                 sec_group = ec2.create_security_group(GroupName='webserver', Description='webserver', VpcId=vpc)
-                log.info("Created Web SG...") 
+                log.info("Created Web SG: " + sec_group.group_id ) 
                 rule1 = sec_group.authorize_ingress( 
                      IpProtocol='tcp',
                      FromPort=80,
@@ -96,11 +96,13 @@ def build(ctx,ami,instancetype,vpc,isweb,subnet,key):
 def configure_web(host,key):
     ssh_user = "ec2-user"
     log.info("Attempting to connect as " + ssh_user)
+    log.info("Using key.." + key)
     try:
         log.info("Waiting for 60 seconds for EC2 instance to finish setting up..")
         time.sleep(60)        
         ssh = paramiko.SSHClient()
-        key = paramiko.RSAKey.from_private_key_file(os.getcwd() +'/' +  key + '.pem')
+        ssh_key = key
+        key = paramiko.RSAKey.from_private_key_file(os.getcwd() +'/' +  str(key) + '.pem')
         ssh.set_missing_host_key_policy(paramiko.AutoAddPolicy())
         connection = ssh.connect(hostname=host, username=ssh_user, pkey=key)
         log.info("Attempting Connection on host: " + host)
@@ -111,7 +113,7 @@ def configure_web(host,key):
             log.info(ssh_stdout)
             log.info("Apache install is Complete!!")
             time.sleep(5)
-            apache_turnon(host,ssh_user,key)
+            apache_turnon(host,ssh_user,ssh_key)
         except Exception as brr:
             log.error(brr)
             sys.exit()
@@ -120,27 +122,39 @@ def configure_web(host,key):
         sys.exit()
 
 @main.command()
-@click.option('--instanceid',help='AWS Instance Id')
+@click.option('--instanceid',required=False,help='AWS Instance Id')
 @click.option('--sgid',required=False,help='Coming Soon... For now, this is Optional')
 @click.pass_context
 def destroy(ctx,instanceid,sgid):
    """ Destroy EC2 Instance """
-   log.info("Terminating Instance...")
-   try:
-       ec2 = connection()
-       instance = ec2.Instance(instanceid)
-       instance.terminate()
-       log.info("Instance ID: " + instanceid + " is destroyed.")
-   except Exception as err:
-       log.error(err)
-       sys.exit() 
+   if instanceid:
+       try:
+           log.info("Terminating Instance...")
+           ec2 = connection()
+           instance = ec2.Instance(instanceid)
+           instance.terminate()
+           log.info("Instance ID: " + instanceid + " is destroyed.")
+       except Exception as err:
+           log.error(err)
+           sys.exit() 
+   if sgid:
+       log.info("Deleting Security group ID: " + sgid)
+       try:
+           client = boto3.client('ec2',aws_access_key_id=access_key,aws_secret_access_key=secret_key,region_name=region)
+           client.delete_security_group(GroupId=sgid)
+           log.info("Security Group has been deleted")
+       except Exception as err:
+           log.error(err)
+           sys.exit()
+       
 
 def apache_turnon(host,ssh_user,key):
-    ## Need to see why I can't str concat on key variable. I am passing ssh_key as the variable - when I tried printing it out, it seemed to print characters out. I will fix this at a later date. 
+    ## Need to see why I can't str concat on key variable. I am passing key as the variable - when I tried printing it out, it seemed to print characters out. I will fix this at a later date. 
+    log.info("Logging in with SSH Key: " + key + " to turn Apache on")
     ssh = paramiko.SSHClient()
-    key = paramiko.RSAKey.from_private_key_file(os.getcwd() + '/test2.pem')
+    key = paramiko.RSAKey.from_private_key_file(os.getcwd() + '/' + key + '.pem')
     ssh.set_missing_host_key_policy(paramiko.AutoAddPolicy())
-    connection = ssh.connect(hostname=host, username=ssh_user, pkey=ssh_key)
+    connection = ssh.connect(hostname=host, username=ssh_user, pkey=key)
     service_on='sudo service httpd start'
     ssh_stdn, ssh_stdout, ssh_stderr=ssh.exec_command(service_on)
     log.info("Try accessing the website: http://" + host + ":80")
